requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
apscheduler==3.10.4
bcrypt==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary>=2.9.9
facebook-business==18.0.4
google-ads>=24.1.0
resend==2.3.0
pytest==8.2.0
pytest-flask==1.3.0
